feat: user role repository and service

+++ internal/core/domain/role.go
@@ -1,6 +1,6 @@
 package domain
 
 type UserRole struct {
-	UserId int64
-	RoleId int64
+	UserId int32
+	RoleId int32
 }
